app/requirements.txt
Django>=4.2.0,<4.3.0
psycopg2-binary>=2.9.6,<3.0.0
gunicorn>=21.2.0,<22.0.0
# Removed non-existent package python-logging-handler
# Use standard Python logging instead
whitenoise>=6.5.0,<7.0.0   # For serving static files
pillow>=10.0.0,<11.0.0     # For image processing
python-dotenv>=1.0.0,<2.0.0  # For environment variables
django-chunked-upload>=2.0.0,<3.0.0  # For handling large file uploads
python-magic>=0.4.27,<0.5.0  # For file type detection
